added addNewFeature to helpers.ts
