+++ elements/map/helpers.ts
@@ -2,6 +2,9 @@
 import GeoJSON from "ol/format/GeoJSON";
 import KML from "ol/format/KML";
 import TopoJSON from "ol/format/TopoJSON";
+import { Vector as VectorLayer } from "ol/layer";
+import { Vector as VectorSource } from "ol/source";
+import { EOxMap } from "./main";
 
 /**
  * Define the read options with proper typing
@@ -12,12 +15,52 @@ const READ_FEATURES_OPTIONS = {
 };
 
 /**
+ * Drag and drop upload shape file's event
+ *
+ * @param {object} e
+ * @param {VectorLayer<VectorSource>} drawLayer
+ * @param {EOxMap} EOxMap
+ */
+export function addNewFeature(
+  e: object,
+  drawLayer: VectorLayer<VectorSource>,
+  EOxMap: EOxMap
+) {
+  const currFeatures = drawLayer.getSource().getFeatures().length;
+  if (
+    !drawLayer.get("multipleFeatures") &&
+    (currFeatures || e.features.length > 1)
+  )
+    throw new Error("Multiple features detected!");
+
+  //@ts-ignore
+  e.features.forEach((feature) => feature.set("id", feature.ol_uid));
+
+  //@ts-ignore
+  drawLayer.getSource().addFeatures(e.features);
+  EOxMap.map
+    .getView()
+    .fit(drawLayer.getSource().getExtent(), { duration: 750 });
+
+  const drawendEvt = new CustomEvent("addfeatures", {
+    detail: {
+      originalEvent: e,
+    },
+  });
+  EOxMap.dispatchEvent(drawendEvt);
+}
+
+/**
  * This function reads text from the clipboard and attempts to parse it as GeoJSON, KML, or TopoJSON.
  * If successful, it adds the parsed features to the map.
  *
- * @param {function} addNewFeature
+ * @param {VectorLayer<VectorSource>} drawLayer
+ * @param {EOxMap} EOxMap
  */
-export function pasteFeaturesFromClipboard(addNewFeature): void {
+export function pasteFeaturesFromClipboard(
+  drawLayer: VectorLayer<VectorSource>,
+  EOxMap: EOxMap
+): void {
   navigator.clipboard
     .readText()
     .then((text: string) => {
@@ -38,7 +81,7 @@ export function pasteFeaturesFromClipboard(addNewFeature): void {
           return;
         }
 
-        addNewFeature({ features });
+        addNewFeature({ features }, drawLayer, EOxMap);
       } catch (err) {
         console.error("Error parsing data from clipboard:", err);
       }
